* elpa-admin.el: Improve doc build for the :core packages
(elpaa--select-revision): Update in the target of symlinks
(elpaa--insert-repolinks): Fix problem for :core packages with files in
completely different directories.
(elpaa--build-Info-1): Add the surrounding dir (including the `emacs`
subdir) to the sandbox, and don't follow symlinks when opening an Org
file so we don't try to output the texi in the `emacs` subtree.