* elpa-admin.el: Fix mime types for org and md
Also use symbols rather than strings for mime types.
And build the temp files in the current directory since that's
the only dir where we have write access when running in the sandbox.

(elpaa--section-to-html, elpaa--export-org):
Build the temp files in the current directory.
(elpaa--extension-to-mime): New function.
(elpaa--get-section): Use it.
(elpaa--get-NEWS): Adjust to new return value of `elpaa--get-section`.
(elpaa--html-make-pkg): Rename some vars.