* elpa-admin.el: Fix computation of list of docs
We can't compute the list of HTML docs from the `:doc` property (and
even if we could it would not be a good idea), so instead keep track
of those docs as we build them so the data is readily available.

(elpaa--html-insert-docs): Get the HTML docs names from the pkg-spec.
(elpaa--html-build-doc): Add `pkg-spec` arg.
Put the doc's info into the `pkg-spec`.
(elpaa--build-Info-1): Add `pkg-spec` arg accordingly.
(elpaa--build-Info): Adjust accordingly.