* elpa-admin.el (elpaa--worktree): Improve output
Do not put output of two different git invocations on the same line.

I had problems locating the code that was responsible for the missing
newline character because of the location of the `message' that prints
"Cloning branch", so I also move that to where I would have expected
it, i.e. before the code where the cloning happens.