* elpa-admin.el (elpaa--make-one-tarball): Ensure auxiliary files exist
Various auxiliary files are created by the functions whose main task
it is to create the tarball.  Previously they didn't do anything if
the tarball exists but one or more of the auxiliary files are missing.
Now they (re)create all the files if any of them are missing.

Previously there was some fallback code to generate just the least
important of the auxiliary files (<pkgname>.svg>) if it was missing.
That doesn't make any sense; either recreate them all or none, but
certainly not only the least important one.