Make used commit and package version known
Expose the information to `make` and `makeinfo` processes, so that
they can use it to, for example, bump version strings in manuals.

Also store the used commit in the "<package>-pkg.el" file to make it
later easier to determine which commit exactly a published package was
built from.

* elpa-admin.el (elpaa--select-revision): If REV argument is a
function, then use its return value.  Return the selected revision.
* elpa-admin.el (elpaa--make-one-tarball-1): Do not evaluate the
function that returns the revision, instead let
`elpaa--select-revision` do it.  Bind `process-environment' around
the code that calls `make` and `makeinfo` to the environment
returned by `elpaa--makeenv`.
* elpa-admin.el (elpaa--makeenv): New function. Manipulate a copy
of process-environment to set PACKAGE_VERSION and PACKAGE_REVISION
to the values passed in as arguments.
* elpa-admin.el (elpaa--write-pkg-file): Include the revision as
specified in the new REVISION argument in the written data.  If
that argument is nil or not specified, then calculate the revision.