Add support for GNU Guix
* elpa-admin.el (elpaa--sandbox-ro-binds): Add /gnu to the list of
read-only directories so that a "Guix" Emacs can access its site-lisp
files.
* elpa-manifest.scm: Add a Guix manifest file with all the necessary
packages to use GNU ELPA.

The manifest file is added to the admin branch, so that if
elpa-admin.el requires a new dependency, the change can be made
centrally and will be reflected in both elpa.git and nongnu.git.

elpa-manifest.scm is symbolically linked into the repositories'
respective root directories at initialisation, as is already the case
with the GNUmakefile.  By linking it, the file will be renamed to the
more conventional "manifest.scm".  We use "elpa-manifest.scm" here to
avoid confusing a "guix shell" invocation, that is hard-coded to check
for a "manifest.scm" and "guix.scm" file.