(elpaa--make-one-tarball-1): Refresh metadata
The metadata passed to `elpaa--make-one-tarball` is read from
the HEAD revision, so it can contain incorrect info (e.g. pre-requisites
that only apply to the not-yet-released code).  Refresh it when needed.

* elpa-admin.el (elpaa--keep-max): Use `defvar`, it's not a constant.
(elpaa--make-one-tarball): Rename `metadata` arg according to its new use.
(elpaa--make-one-tarball-1): Rename `metadata` arg and recompute the
metadata if requested after we selected the revision.
(elpaa--make-one-package): Tell `elpaa--make-one-tarball-1` to
recompute the metadata for the release tarballs.