* elpa-admin.el: Correctly restore the original branch for `:core`
The code to save&restore the worktree/branch state was not run in the
proper directory for `:core` and that part that did had not been adjusted
to the use of a "detached checkout".  Furthermore it was partly
duplicated in `elpaa--clean` and `elpaa--select-revision`.

(elpaa--clean): Move the code that saves&restores the
original revision from here....
(elpaa--select-revision): ...to here.  Tweak it along the way so it runs
in the same directory as the code that does the `checkout` of the selected
commit, and don't bother with it if we don't select another commit.