Improve support for "subpackages"
Try and provide a bit more support for packages built from the same
branch, such as helm/helm-core.  The idea is to try and keep a single
local `elpa/<FOO>` branch (and a single local worktree) for all the
packages coming from the same upstream branch.

* elpa-admin.el (elpaa--get-specs): Turn `:url <PKGNAME>` into
the usual `:url URL` format (and remember the parent in `:parent--package`).
(elpaa--local-branch-name): New function.
(elpaa--get-last-release, elpaa--check-sync-failures, elpaa--push)
(elpaa--publish-package-spec, elpaa--insert-repolinks, elpaa--ortb): Use it.
(elpaa--worktree-sync): Use it.  Also, use a symlink for subpackages.
(elpaa--batch-fetch-and): Skip subpackages.